refactor: shared dialog css, update containers to normalize dialog appearance
